+++ waves/tests/test_parameter_study.py
@@ -1,8 +1,10 @@
 import io
 import pathlib
-from unittest.mock import patch, mock_open
+from unittest.mock import Mock, patch, mock_open
+from contextlib import nullcontext as does_not_raise
 
 import pytest
+import yaml
 
 from waves import _parameter_study
 
@@ -38,3 +40,31 @@ def test_read_parameter_schema():
          pytest.raises(RuntimeError):
         parameter_schema = _parameter_study.read_parameter_schema(None)
     mock_file.assert_not_called()
+
+
+def test_main():
+    # Check the YAML read error and clarification runtime error message
+    with patch("waves._parameter_study.read_parameter_schema", side_effect=yaml.parser.ParserError()), \
+         patch("waves.parameter_generators.ParameterGenerator") as mock_class, \
+         pytest.raises(RuntimeError):
+        try:
+            _parameter_study.main("cartesian_product", "dummy.yaml")
+        finally:
+            mock_class.assert_not_called()
+
+
+    # Check correct subcommand/ParameterGenerator class associations
+    associations = (
+        ("cartesian_product", "CartesianProduct"),
+        ("custom_study", "CustomStudy"),
+        ("latin_hypercube", "LatinHypercube"),
+        ("sobol_sequence", "SobolSequence"),
+    )
+    for subcommand, generator in associations:
+        mock_generator = Mock()
+        with patch("waves._parameter_study.read_parameter_schema", return_value={}), \
+             patch(f"waves.parameter_generators.{generator}", return_value=mock_generator) as mock_class, \
+             does_not_raise():
+            _parameter_study.main(subcommand, "dummy.yaml")
+            mock_class.assert_called_once()
+            mock_generator.write.assert_called_once()